vectordb refactoring, will fix #72
